+++ ChangeLog.md
@@ -6,6 +6,9 @@ All notable changes to gama-q2 project are documented in this file.
 
 Build on Ubuntu 20.04.1 LTS (qtcreator and cmake -GNinja)
 
+* added missing 'emit' keywords and fixed inserting point in
+  PointTableModel::insertRows (observationtablemodel.cpp)
+
 * fixed warnings connect-3arg-lambda
 
 * fixed compile warnings "c++11 range-loop might detach Qt container"